+++ x-pack/plugins/observability/public/hooks/slo/use_fetch_historical_summary.ts
@@ -9,7 +9,7 @@ import { useQuery } from '@tanstack/react-query';
 import { FetchHistoricalSummaryResponse } from '@kbn/slo-schema';
 
 import { useKibana } from '../../utils/kibana_react';
-import { sloKeys } from './query-key-factory';
+import { sloKeys } from './query_key_factory';
 
 export interface UseFetchHistoricalSummaryResponse {
   sloHistoricalSummaryResponse: FetchHistoricalSummaryResponse | undefined;
