+++ x-pack/plugins/observability/public/hooks/slo/use_fetch_rules_for_slo.ts
@@ -13,7 +13,7 @@ import {
 } from '@tanstack/react-query';
 import type { Rule } from '@kbn/triggers-actions-ui-plugin/public';
 import { useKibana } from '../../utils/kibana_react';
-import { sloKeys } from './query-key-factory';
+import { sloKeys } from './query_key_factory';
 
 type SloId = string;
