+++ x-pack/plugins/observability/public/hooks/slo/use_fetch_slo_list.ts
@@ -17,7 +17,7 @@ import { i18n } from '@kbn/i18n';
 import { FindSLOResponse } from '@kbn/slo-schema';
 
 import { useKibana } from '../../utils/kibana_react';
-import { sloKeys } from './query-key-factory';
+import { sloKeys } from './query_key_factory';
 
 interface SLOListParams {
   name?: string;
